requirements/base.txt
fastapi==0.87.0
SQLAlchemy==1.4.44
uvicorn==0.20.0
Jinja2==3.1.2
python-dotenv==0.21.0
requests==2.28.1
python-jose==3.3.0
python-multipart==0.0.5
databases==0.6.1
aiosqlite==0.17.0
fastapi-limiter==0.1.5
aioredis==2.0.1
loguru==0.6.0
psycopg2-binary==2.9.3
asyncpg==0.27.0
pydantic[email]
passlib[bcrypt]
databases[sqlite]
fastapi[all]
